requirements.txt
numpy==1.14.3
pandas==0.23.4
Pillow==5.2.0
PyPDF2==1.26.0
tabula-py==1.3.1
termcolor==1.1.0
tesseract==0.1.3
ujson==1.35
urllib3==1.23
xlrd==1.0.0
XlsxWriter==1.1.0